Install QEMU emulator in arm64 variant of cpubuilder image. (#13)
Follow-up to https://github.com/iree-org/base-docker-images/pull/11.

This emulator was originally installed in
https://github.com/iree-org/iree/pull/16331. We've been carrying it
around since then.

I'm not thrilled with the file sitting in a cloud storage bucket (GCS or
Azure). The file is 5MB so maybe we could check it in here via git LFS?
Or we could build it from source via our automation.